Refactor to change package name
The change from `com.farbod` to `com.satsuware` is due to the fact that I did not own the domain 'farbod.com', so it would be better to use a domain that I do own. The domain 'satsuware.com' (also (and short for) 'satsumasoftware.com') will be used instead.

To signify these changes, the version has been incremented to 1.0 (then to 1.0.1 to fix some minor issues).

Furthermore, the Android Support library dependencies in `build.gradle` have been updated to their latest versions.